move client_settings_recv to pymine/logic/handle/play
